removal of functions that supports DLL payloads
