+++ winpwnage/functions/elevate/elevateMethod3.py
@@ -1,6 +1,8 @@
 from winpwnage.core.prints import *
 from winpwnage.core.utils import *
 from winpwnage.core.winstructures import *
+import sys
+import os
 
 #Creds to: https://gist.github.com/highsenburger69/09b816daa16f020d188c289fd401b0b2
 
@@ -12,6 +14,7 @@ elevateMethod3_info = {
 	"Fixed In": "99999" if information().admin() == True else "0",
 	"Works From": "7600",
 	"Admin": True,
+	"Syntax": os.path.basename(sys.argv[0]) + " --use elevate --id 3 --payload <Fully qualified path>",
 	"Function Name": "elevateMethod3",
 	"Function Payload": True,
 }
