+++ winpwnage/functions/elevate/elevateMethod5.py
@@ -1,7 +1,9 @@
-import time
 from winpwnage.core.prints import *
 from winpwnage.core.utils import *
 from winpwnage.core.winstructures import *
+import time
+import sys
+import os
 
 elevateMethod5_info = {
 	"Description": "Elevate from administrator to NT AUTHORITY SYSTEM using wmic.exe (non interactive)",
@@ -11,6 +13,7 @@ elevateMethod5_info = {
 	"Fixed In": "99999" if information().admin() == True else "0",
 	"Works From": "7600",
 	"Admin": True,
+	"Syntax": os.path.basename(sys.argv[0]) + " --use elevate --id 5 --payload <Fully qualified path>",
 	"Function Name": "elevateMethod5",
 	"Function Payload": True,
 }
