+++ winpwnage/functions/elevate/elevateMethod6.py
@@ -1,6 +1,7 @@
 from winpwnage.core.prints import *
 from winpwnage.core.utils import *
 from winpwnage.core.winstructures import *
+import sys
 import os
 
 elevateMethod6_info = {
@@ -11,6 +12,7 @@ elevateMethod6_info = {
 	"Fixed In": "99999" if information().admin() == True else "0",
 	"Works From": "7600",
 	"Admin": True,
+	"Syntax": os.path.basename(sys.argv[0]) + " --use elevate --id 6 --payload <Fully qualified path>",
 	"Function Name": "elevateMethod6",
 	"Function Payload": True,
 }
