+++ winpwnage/functions/uac/uacMethod5.py
@@ -4,7 +4,7 @@ from winpwnage.core.prints import *
 from winpwnage.core.utils import *
 
 uacMethod5_info = {
-	"Description": "UAC bypass using sdclt.exe (isolatedcommand)",
+	"Description": "UAC bypass using sdclt.exe (IsolatedCommand)",
 	"Method": "Method: Registry key (Class) manipulation",	
 	"Id": "5",
 	"Type": "UAC bypass",
